requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-telegram-bot[webhooks]==22.0
yt-dlp==2026.03.17
